Resolved merge conflics with Pam4.0 branch
Conflicts were resolved and updated to work with the backend changes I
made.
