--- StudyRingMeteor/client/rings/RingDash.html
@@ -1,29 +1,0 @@
-<!--Ring Dashboard will contain future features such as an organized display of
-    events, members, forum or communication between members and other information-->
-
-<template name="RingDash">
-    <article class="ring">
-        <h3>
-        {{ring.name}}
-        </h3>
-        <br/>
-            Description: {{ring.desc}}
-        <br/>
-            Tags:
-            {{#each ring.tags}}
-                <span class="tags">{{name}}</span>
-            {{/each}}
-        <br/>
-        <p>
-            Members:
-            {{#each members}}
-                <span class="names">{{username}}
-                                    {{> avatar userId=_id shape="circle" size="medium"}}
-                </span>
-            {{/each}}
-        </p>
-
-        Created By: {{createdByUser.username}} {{> avatar userId=createdByUser shape="circle" size="medium"}}
-        {{> NewEvent}}
-    </article>
-</template>
