--- StudyRingMeteor/client/rings/RingDash.js
@@ -1,27 +1,0 @@
-import { Rings } from "../../collections/rings/rings.js";
-
-Template.RingDash.onCreated(function(){
-    this.editMode = new ReactiveVar(false);
-    var self = this;
-    self.autorun(function (){
-        var id = FlowRouter.getParam('id');
-        self.subscribe('singleRing', id);
-        self.subscribe('allUsers');
-    });
-});
-
-Template.RingDash.helpers({
-    ring: ()=> {
-        var id = FlowRouter.getParam('id');
-        return Rings.findOne({_id: id});
-    },
-    createdByUser: function() {
-        var id = FlowRouter.getParam('id');
-        var userid = Rings.findOne({_id: id}).createdBy;
-        return Meteor.users.findOne({_id: userid});
-    },
-    members: () => {
-        var id = FlowRouter.getParam('id');
-        return Meteor.users.find({rings: id});
-    }
-});
