--- StudyRingMeteor/client/rings/RingInfo.html
@@ -1,31 +1,0 @@
-<template name="RingInfo">
-    <article class="ring {{#if isPrivate}}is-private{{/if}}">
-        <h3>
-            Name: {{name}}
-        </h3>
-        {{#if editMode}}
-                {{> quickForm collection=Rings id=updateRingId type="update" doc=this autosave=true}}
-                {{#if isPrivate}}
-                    <button class="btn-primary set-public">Set Public</button>
-                {{else}}
-                    <button class="btn-deny set-private">Set Private</button>
-                {{/if}}
-            {{else}}
-            <p>
-                Description: {{desc}}
-            </p>
-            <p>
-                Tags:
-                {{#each tags}}
-                    <span class="tags">{{name}}</span>
-                {{/each}}
-            </p>
-            <p>
-                Created By: {{user.username}} {{> avatar userId=userId shape="circle" size="medium"}}
-            </p>
-            <a href="/rings/{{_id}}">View Details</a>
-        {{/if}}
-        <i class="fa fa-pencil"></i>
-        <i class="fa fa-trash"></i>
-    </article>
-</template>
