--- StudyRingMeteor/client/rings/RingSingle.html
@@ -1,29 +1,0 @@
-<template name="RingSingle">
-    <article class="ring">
-        <h3>
-            Name: {{name}}
-        </h3>
-        <p>
-            Description: {{desc}}
-        </p>
-        <p>
-            Tags:
-            {{#each tags}}
-                <span class="tags">{{name}}</span>
-            {{/each}}
-        </p>
-        <p>
-            Members:
-            {{#each members}}
-                <span class="names">{{username}}
-                                    {{> avatar userId=_id shape="circle" size="medium"}}
-                </span>
-            {{/each}}
-        </p>
-        {{#if isMember}}
-            <button class="btn-deny leave-ring">Leave Ring</button>
-        {{else}}
-            <button class="btn-primary join-ring">Join Ring</button>
-        {{/if}}
-    </article>
-</template>
